e2e: refactor utils into e2eutil/ pkg
