+++ test/e2e/upgrade_test.go
@@ -18,6 +18,7 @@ import (
 	"testing"
 
 	"github.com/coreos/etcd-operator/pkg/util/k8sutil"
+	"github.com/coreos/etcd-operator/test/e2e/e2eutil"
 	"github.com/coreos/etcd-operator/test/e2e/framework"
 
 	"k8s.io/kubernetes/pkg/api"
@@ -25,33 +26,33 @@ import (
 
 func TestEtcdUpgrade(t *testing.T) {
 	f := framework.Global
-	origEtcd := makeEtcdCluster("test-etcd-", 3)
-	origEtcd = etcdClusterWithVersion(origEtcd, "v3.0.12")
-	testEtcd, err := createEtcdCluster(f, origEtcd)
+	origEtcd := e2eutil.MakeEtcdCluster("test-etcd-", 3)
+	origEtcd = e2eutil.EtcdClusterWithVersion(origEtcd, "v3.0.12")
+	testEtcd, err := e2eutil.CreateEtcdCluster(f, origEtcd)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	defer func() {
-		if err := deleteEtcdCluster(f, testEtcd.Name); err != nil {
+		if err := e2eutil.DeleteEtcdCluster(f, testEtcd.Name); err != nil {
 			t.Fatal(err)
 		}
 	}()
 
-	_, err = waitSizeReachedWithFilter(f, testEtcd.Name, 3, 60, func(pod *api.Pod) bool {
+	_, err = e2eutil.WaitSizeReachedWithFilter(f, testEtcd.Name, 3, 60, func(pod *api.Pod) bool {
 		return k8sutil.GetEtcdVersion(pod) == "v3.0.12"
 	})
 	if err != nil {
 		t.Fatalf("failed to create 3 members etcd cluster: %v", err)
 	}
 
-	testEtcd = etcdClusterWithVersion(testEtcd, "v3.1.0-alpha.1")
+	testEtcd = e2eutil.EtcdClusterWithVersion(testEtcd, "v3.1.0-alpha.1")
 
-	if err := updateEtcdCluster(f, testEtcd); err != nil {
+	if err := e2eutil.UpdateEtcdCluster(f, testEtcd); err != nil {
 		t.Fatalf("fail to update cluster version: %v", err)
 	}
 
-	_, err = waitSizeReachedWithFilter(f, testEtcd.Name, 3, 60, func(pod *api.Pod) bool {
+	_, err = e2eutil.WaitSizeReachedWithFilter(f, testEtcd.Name, 3, 60, func(pod *api.Pod) bool {
 		return k8sutil.GetEtcdVersion(pod) == "v3.1.0-alpha.1"
 	})
 	if err != nil {
